requirements.txt
torch==2.2.0
transformers==4.56.0
tqdm==4.67.1
Pillow==9.5.0
numpy==2.0.1
